requirements.txt
numpy
tabula-py